backend/requirements.txt
fastapi==0.125.0
uvicorn==0.40.0
tavily-python==0.7.21
groq==1.0.0
python-dotenv==1.2.1
aiosqlite==0.22.1
pydantic==2.12.5
google-generativeai>=0.8.0
fpdf2>=2.7.0
chromadb>=0.4.0